Rock Paper Scissors.py
A fellow Python coder here. Just wanted to add a file that highly corresponds to the files/folders in your repository. I added a fully working game of Rock Paper and Scissors. Hope it helps and you can check and then merge the request. Thank you,
P.S.- It's very easy to maintain and has variable_name.capitalize() to capitalize the first letter of the inputted word so "Rock" and "rock" would yield similar result instead of creating an error.